Rename webpack baseConfig.js
